requirements.txt
google-generativeai==0.8.3
requests==2.32.3
beautifulsoup4==4.12.3
PyMuPDF==1.24.11
pandas==2.2.3
python-dotenv==1.0.1
tenacity==9.0.0
aiohttp==3.9.5